backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
Flask-Mail==0.9.1
twilio==8.10.0
stripe==7.3.0
bcrypt==4.0.1
Pillow==10.0.0
python-dateutil==2.8.2